Added OpenStreet Maps Test page, Updated 360 iframe to bootstrap responsive
In this commit I added the gorumna-island.hbs file to test openstreet maps. To make this work as another page
I added another route to the index.js file making use of the express.js library.
I added divs around the existing 360 degree image on the main page to make it responsive and added a button
to this new page.