Map generated corresponding to the location
Map is now being generated correctly matching with the place, works by reading the div being displayed, also has a marker that is centered in the map with the name of location and its difficulty. The marker will be able to be placed elsewhere if needed using marker coordinates. Will need to change the implementation if not displaying coordinates