/addPlace working with form and added audio
This means that there are 3 images and 3 audio sounds to choose from for a place.
They are chosen and send to the /addPlace api which is now working and that data is sent to the mongodb database.
I also fixed another error with map.js whereby it searched for grade which no longer existed.